refactor: base64toUint8 moved to utils directory
