+++ apps/dashboard/pages/index.tsx
@@ -14,25 +14,12 @@ import {
   StrapiLocale,
 } from '@fc/types'
 import { AdminLayout } from '@fc/ui'
+import { base64ToUint8Array } from '@fc/utils'
 
 const args: RequestCollectionArgs<AccountStats> = {
   endpoint: 'account-statistics',
   sort: ['date:asc'],
   pageSize: 100,
-}
-
-const base64ToUint8Array = (base64: string) => {
-  const padding = '='.repeat((4 - (base64.length % 4)) % 4)
-  const b64 = (base64 + padding).replace(/-/g, '+').replace(/_/g, '/')
-
-  const rawData = window.atob(b64)
-  const outputArray = new Uint8Array(rawData.length)
-
-  for (let i = 0; i < rawData.length; ++i) {
-    outputArray[i] = rawData.charCodeAt(i)
-  }
-
-  return outputArray
 }
 
 const Index = () => {
