+++ package.json
@@ -140,9 +140,9 @@
     "storybook": "^7.6.19",
     "stream-browserify": "^3.0.0",
     "turbo": "^1.13.3",
-    "webpack": "^5.91.0",
     "type-fest": "^4.18.2",
-    "typescript": "^5.4.5"
+    "typescript": "^5.4.5",
+    "webpack": "^5.91.0"
   },
   "packageManager": "yarn@4.2.1",
   "workspaces": [
