+++ packages/utils/src/index.ts
@@ -15,5 +15,6 @@ export * from './mapModelLocalization'
 export * from './mapSessionUser'
 export * from './sleep'
 export * from './toBase64'
+export * from './toUint8Array'
 export * from './toastMessage'
 export * from './permissions'
