+++ packages/utils/src/toUint8Array.ts
@@ -1,0 +1,13 @@
+export const base64ToUint8Array = (base64: string) => {
+    const padding = '='.repeat((4 - (base64.length % 4)) % 4)
+    const b64 = (base64 + padding).replace(/-/g, '+').replace(/_/g, '/')
+  
+    const rawData = window.atob(b64)
+    const outputArray = new Uint8Array(rawData.length)
+  
+    for (let i = 0; i < rawData.length; ++i) {
+      outputArray[i] = rawData.charCodeAt(i)
+    }
+  
+    return outputArray
+  }
